python_mysql/python/src/requirements.txt
alembic==1.11.1
arrow==1.2.3
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.1.1
packaging==23.2
pycountry==22.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4